==> notebooks/02_train_and_evaluate_no_nigeria_training.ipynb <==
import sys
from argparse import ArgumentParser, Namespace

sys.path.append("..")

from src.models import STR2MODEL, train_model

## 1. Initialize model 

parser = ArgumentParser()
parser.add_argument("--max_epochs", type=int, default=1000)
parser.add_argument("--patience", type=int, default=10)

**Slight modification below:** adding args=[] makes it work for Jupyter notebook https://stackoverflow.com/questions/30656777/how-to-call-module-written-with-argparse-in-ipython-notebook

model_args = STR2MODEL["land_cover"].add_model_specific_args(parser).parse_args(args=[])

model = STR2MODEL["land_cover"](model_args)
model.hparams

### New models args

new_model_args_dict = vars(model_args)

new_model_args_dict

# SET MODIFICATIONS TO DEFAULT MODEL ARGUMENTS:

new_model_args_dict['add_togo'] = False
new_model_args_dict['multi_headed'] = False
new_model_args_dict['num_classification_layers'] = 1
new_model_args_dict['max_epochs'] = 1 # Just for dev
new_model_args_dict['gpus'] = 0
new_model_args_dict['remove_b1_b10'] = False

# Initialize model with new arguments
new_model_args = Namespace(**new_model_args_dict)
model = STR2MODEL["land_cover"](new_model_args)
model.hparams

## 2. Check data class distribution

from collections import Counter

### Training set

loader = model.train_dataloader()
train_counter = Counter()
for sample in loader:
    x, y, weight = sample
    train_counter.update(y.numpy())

print('Non-croplad / Cropland in training set (%):')
a = train_counter[0.0] / (train_counter[0.0] + train_counter[1.0])
a, 1-a

### Validation set

loader = model.val_dataloader()
val_counter = Counter()
for sample in loader:
    x, y, weight = sample
    val_counter.update(y.numpy())

val_counter
print('Non-croplad / Cropland in validation set (%):')
a = val_counter[0.0] / (val_counter[0.0] + val_counter[1.0])
a, 1-a

### Test set

loader = model.test_dataloader()
test_counter = Counter()
for sample in loader:
    x, y, weight = sample
    test_counter.update(y.numpy())

test_counter
print('Non-croplad / Cropland in test set (%):')
a = test_counter[0.0] / (test_counter[0.0] + test_counter[1.0])
a, 1-a

## 3. Train model 

train_model(model, new_model_args)
