==> notebooks/03_run_nigeria_export.ipynb <==
import sys
from pathlib import Path


sys.path.append("..") 

from src.exporters import NigeriaSentinelExporter, GDriveExporter, cancel_all_tasks

Before I have run `process_nigeria()` in `scripts/process.py`

def export_nigeria():
    exporter = NigeriaSentinelExporter(Path("../data"))
    exporter.export_for_labels(
        num_labelled_points=None, monitor=False, checkpoint=True
    )

export_nigeria()

#cancel_all_tasks()

def export_gdrive_nigeria():
    exporter = GDriveExporter(Path('../data'), dataset=NigeriaSentinelExporter.dataset)
    exporter.export()

export_gdrive_nigeria()
