==> notebooks/07_visual_classes_comparison_for_labelling.ipynb <==
import pathlib
from pathlib import Path

import pandas as pd
import geopandas as gpd

## Nigeria and togo boundaries

in_path = Path('/home/gajo/DFS/Projects/2021-data-org/4._RESEARCH_n/ML/MLsatellite/data/layers-general/administrative_borders/ne_50m_admin_0_countries/ne_50m_admin_0_countries.shp')
out_folder_path = Path('/home/gajo/DFS/Projects/2021-data-org/4._RESEARCH_n/ML/MLsatellite/Research/WP1_Joaquin/paper-nigeria-cropland/layers')

for country in ['Nigeria', 'Togo']:
    world = gpd.read_file(in_path)
    gdf = world[world['SOVEREIGNT'] ==  country].reset_index(drop=True)
    out_path = out_folder_path / f'{country.lower()}_borders.shp'
    if not out_path.exists():
        gdf.to_file(out_folder_path / f'{country.lower()}_borders.shp')
    else:
        print(out_path, 'already exists!')

## Labels

togo_train_path = Path('../data/processed/togo/data.geojson')
togo_eval_path = Path('../data/processed/togo_evaluation/data.csv')

df_train = gpd.read_file(togo_train_path)
df_train

df_test = pd.read_csv(togo_eval_path)
df_test
