==> notebooks/09_nigeria_new_data_export.ipynb <==
import sys
from pathlib import Path

import ee

sys.path.append("..") 

from src.exporters import NigeriaSentinelExporterNew, GDriveExporter, cancel_all_tasks
from src.processors import NigeriaProcessorNew
from src.engineer import NigeriaEngineerNew
from src.utils.samplers import BufferedResampler

%reload_ext autoreload
%autoreload 2

#ee.Authenticate()

processor = NigeriaProcessorNew(Path("../data"))
processor.process()

exporter = NigeriaSentinelExporterNew(Path("../data"))
exporter.export_for_labels(
    num_labelled_points=None, monitor=False, checkpoint=True
)

# exporter = GDriveExporter(Path('../data'), dataset=NigeriaSentinelExporterNew.dataset)
# exporter.export()

# Only need to run once
engineer = NigeriaEngineerNew(Path("../data"))
#engineer.engineer(val_set_size=0.2, test_set_size=0.2, calculate_normalizing_dict=True)

resampler = BufferedResampler(data_path=engineer.savedir, test_set_size=0.25, val_set_size=0.25, buffer=30, engineer=engineer)
resampler.resample()
