==> notebooks/13_cropharvest_sentinel1.ipynb <==
from pathlib import Path
from datetime import date

import geopandas as gpd
from cropharvest.eo import EarthEngineExporter

## Export Nigeria labels

#ee.Authenticate()

google_drive_exporter = EarthEngineExporter()

path = Path('../data/processed/nigeria/data.geojson')
gdf = gpd.read_file(path)
gdf

start_date = date(2019, 4, 3)
end_date = date(2020, 3, 28)
gdf['start_date'] = start_date
gdf['end_date'] = end_date
gdf['export_identifier'] = gdf.apply(lambda row: f"{str(row.loc['index'])}_{str(row.loc['start_date'])}_{str(row.loc['end_date'])}", axis=1) # TODO: put '_nigeria_' after index
gdf

#google_drive_exporter.export_for_labels(labels=gdf)

## Train model with geowiki and nigeria data

from cropharvest.datasets import CropHarvest, CropHarvestLabels
from cropharvest.engineer import Engineer
from cropharvest.bands import BANDS, DYNAMIC_BANDS, STATIC_BANDS, REMOVED_BANDS

gdf

path

tifs_folder_path = Path('../data/raw/earth_engine_nigeria_S1/')
tifs_path = list(tifs_folder_path.glob('*.tif'))

tifs_path[0]

import rasterio

with rasterio.open(tifs_path[0], 'r') as f:
    meta = f.meta
    data = f.read()

meta

BANDS
