==> notebooks/14_cropharvest_sentinel1_engineer_nigeria.ipynb <==
from pathlib import Path
from datetime import date

import geopandas as gdp

from cropharvest.datasets import CropHarvest, CropHarvestLabels
from cropharvest.engineer import Engineer
from cropharvest.bands import BANDS, DYNAMIC_BANDS, STATIC_BANDS, REMOVED_BANDS
from cropharvest.columns import RequiredColumns, NullableColumns

data_root = Path('../data/features/nigeria-cropharvest')
file_name = 'labels.geojson'

gdf = gdp.read_file('../data/processed/nigeria/data.geojson')
start_date = date(2019, 4, 3)
end_date = date(2020, 3, 28)
shapefile = gdp.read_file('../data/features/nigeria-cropharvest/nigeria_stratified_labelled_v1_splits_resampled.shp')
shapefile['identifier'] = shapefile['identifier'].astype(int)
gdf = gdf.merge(shapefile[['identifier', 'new_set']], left_on='index', right_on='identifier', suffixes=(None, '_y'))

#gdf['export_identifier'] = gdf.apply(lambda row: f"{str(row.loc['index'])}_{str(row.loc['start_date'])}_{str(row.loc['end_date'])}", axis=1)
gdf['start_date'] = start_date # required by EarthEngineExporter
gdf['end_date'] = end_date # required by EarthEngineExporter
gdf['export_end_date'] = end_date # required by Engineer
gdf['dataset'] = 'nigeria'
gdf['export_identifier'] = gdf.apply(lambda row: f"{str(row.loc['index'])}_{str(row.loc['dataset'])}_{str(row.loc['start_date'])}_{str(row.loc['end_date'])}", axis=1)
gdf['is_test'] = gdf['new_set'] == 'testing'
gdf['label'] = None
gdf

# Save labels 
gdf['start_date'] = gdf['start_date'].astype(str)
gdf['end_date'] = gdf['end_date'].astype(str)
gdf['export_end_date'] = gdf['export_end_date'].astype(str)
gdf.to_file(data_root / file_name, driver='GeoJSON')

engineer = Engineer(data_root)

## Modify file name so that Engineer.process_filename() works
# folder = Path('data/nigeria/eo_data/')
# files = list(folder.glob('*.tif')) 
# for file_path in files:
#     parts = file_path.name.split('_')
#     new_file_name = '_'.join([str(parts[0]) + '-' + 'nigeria, *parts[1:]])
#     file_path.rename(folder / new_file_name)

engineer.create_h5_dataset()
