==> notebooks/16_cropharvest_sentinel1_export_region_nigeria.ipynb <==
import sys
from datetime import date
from pathlib import Path

import ee
import gdown
import pandas as pd
from cropharvest.eo import EarthEngineExporter
from cropharvest.countries import BBox

sys.path.append('..')
from src.exporters import GDriveExporter

#ee.Authenticate()

google_drive_exporter = EarthEngineExporter()

# By default data will be exported to Google Drive
google_drive_exporter.export_for_bbox(
    bbox=BBox(min_lon=2.6917, max_lon=14.5772, min_lat=4.2406, max_lat=13.8659), # https://gist.github.com/graydon/11198540 in W, S, E, N aprox. to 4 decimal
    bbox_name="Nigeria",
    start_date=date(2019, 4, 3),
    end_date=date(2020, 3, 28),
    metres_per_polygon=20000,
)

# NOTE: Since 3000 is the max of tasks GEE can handle simulanteously and we have 3563 tasks int total I added a small patch to the source code of export_for_bbox() at cropharvest.eo.EarthEngineExplorer:
# ...
# for identifier, region in zip(ids, regions):
#             if int(identifier.split('_')[1].split('/')[0]) > 3020:   --> ADDITION
#                 return_obj[identifier] = self._export_for_polygon(
#                     polygon=region,
#                     polygon_identifier=f"{bbox_name}/{identifier}",
#                     start_date=start_date,
#                     end_date=end_date,
#                     file_dimensions=file_dimensions,
#                     test=True,
#                 )
# ...

## Experimenting with how to download from google drive to local (see scripts/export.py now for all files!)

exporter = GDriveExporter(Path('/media/Elements-12TB/satellite_images/nigeria'), dataset='nigeria-full-country-2020')
#file_info = exporter.list_files_in_folder(folder_name='test_eo_data') 
#df = pd.DataFrame(file_info)
#df.to_csv('../gdrive_test_eo_data_files.csv', index=False)

path = Path('../gdrive_test_eo_data_files.csv')
file_info = []
if path.exists():
    print('Loading csv files with names')
    df = pd.read_csv(path)
    file_info = df.to_dict('records')
else:
    print('Listing files on gdrive...')
    file_info = exporter.list_files_in_folder(folder_name='test_eo_data')

file_info

len(file_info)

#browser.get(link)
#df = pd.read_html('https://drive.google.com/uc?id=1X-J4Vc6BFc3MCvQ3SNnW-8Ny_0f_rdTV')

from bs4 import BeautifulSoup

url = 'https://drive.google.com/uc?id=1X-J4Vc6BFc3MCvQ3SNnW-8Ny_0f_rdTV'

import requests

r = requests.get(url)
r.text

len(r.text)

soup = BeautifulSoup(r.text)

soup.find(id='download-form')#.get('action')

file_id = '1X-J4Vc6BFc3MCvQ3SNnW-8Ny_0f_rdTV'
file_id = 'fM2aC0J8lDCKFqIT-Cy6n6GbCml1WcdK'
api_key = 'AIzaSyDOmvYxORH-lB3dV6MN0J9pDU6PhSTpToM'
url = f'https://www.googleapis.com/drive/v3/files/{file_id}?alt=media&key={api_key}'
url

Daily usage limit for this file!

import os
os.system(f"wget {url} -OutFile")

file_id = '1VbB3HYP5TFNMidkeX5195xmmnych03Ud'
file_id = '1YN9ds6LhSJUNYrMB1E6dNtknpT7LPHBS'
url = f'https://www.googleapis.com/drive/v3/files/{file_id}?alt=media&key={api_key}'
url

name = 'dummy_file.txt'
path = Path('/media/Elements-12TB/satellite_images/nigeria/raw') / name
response = os.system(f"curl -o {path} '{url}'")


response

for i, file in enumerate(file_info):
    file_id = file['id']
    file_name = file['name']
    output_path = exporter.output_folder / file_name.split('/')[-1]
    gdown.download(id=file_id, output=output_path, quiet=False)

file_id = '1p2kUd4U4304v0kEQRjG5C4MgFefc1m8y' #'1bzWtQTtpeJhD4h8HwbkGvprPOLya8VVu'
file_name = 'Nigeria/batch_3415/3415-0000001792-0000001792.tif'
output_path = exporter.output_folder / file_name.split('/')[-1]
gdown.download(id=file_id, output=str(output_path), quiet=False)

gdown.download(id=file_info[0]['id'], output=file_info[0]['name'].split('/')[-1], quiet=False)

str(output_path)

## Export from gdrive

exporter = GDriveExporter(Path('/media/Elements-12TB/satellite_images/nigeria'), dataset='nigeria-full-country-2020')

exporter.list_files_in_folder('test_eo_data')

exporter.export('test_eo_data')

Finally works!! need to put credentials.json and token.json in output folder.
