==> notebooks/19_cropharvest_sentinel1_reimplementation RF_benchmark.ipynb <==
from argparse import ArgumentParser, Namespace
import h5py
from itertools import permutations
from pathlib import Path
from typing import cast, Optional, List, Tuple, Dict, Type, TypeVar, Sequence
from tqdm import tqdm
import sys

import numpy as np
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
import torch
from torch.utils.data import Dataset
from shapely.geometry import Point
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score, precision_score, recall_score, roc_auc_score, accuracy_score
from cropharvest.datasets import CropHarvest, CropHarvestLabels, Task
from cropharvest.columns import NullableColumns, RequiredColumns
from cropharvest.config import FEATURES_DIR
from cropharvest.engineer import Engineer
from cropharvest.utils import load_normalizing_dict
from cropharvest.bands import BANDS, DYNAMIC_BANDS, STATIC_BANDS, REMOVED_BANDS

sys.path.append("..")

from src.models import STR2MODEL, STR2BASE, train_model
from src.models.data import LandTypeClassificationDataset, NigeriaCropHarvestDataset, GeowikiCropHarvestDataset

S2_BANDS = ['B2','B3','B4','B5','B6','B7','B8','B8A','B9','B11','B12','NDVI']

DATA_DIR = Path("../data")

# Helper functions
def get_metrics(y_true, y_probs):
    roc_auc = roc_auc_score(y_true, y_probs)
    y_pred = (y_probs > 0.5).astype(int)
    precision = precision_score(y_true, y_pred)
    recall = recall_score(y_true, y_pred)
    f1 = f1_score(y_true, y_pred)
    acc = accuracy_score(y_true, y_pred)

    print('RF roc-auc test set:', round(roc_auc, 3))
    print('RF precision test set:', round(precision, 3))
    print('RF recall test set:', round(recall, 3))
    print('RF f1-score test set:', round(f1, 3))
    print('RF accuracy test set:', round(acc, 3))

    return {'roc_auc': roc_auc, 'precision': precision, 'recall': recall, 'f1': f1, 'acc': acc}

def get_model(add_geowiki: bool, add_nigeria: bool, geowiki_subset: str):
    parser = ArgumentParser()
    parser.add_argument("--max_epochs", type=int, default=100)
    parser.add_argument("--patience", type=int, default=10)
    parser.add_argument("--gpus", type=int, default=0)
    parser.add_argument("--wandb", default=False, action="store_true")

    model_args = STR2MODEL["land_cover"].add_model_specific_args(parser).parse_args(args=[])
    new_model_args_dict = vars(model_args)

    # SET MODIFICATIONS TO DEFAULT MODEL ARGUMENTS:
    new_model_args_dict['add_geowiki'] = add_geowiki
    new_model_args_dict['add_nigeria'] = add_nigeria
    new_model_args_dict['geowiki_subset'] =  geowiki_subset # 'nigeria', 'neighbours1'

    new_model_args = Namespace(**new_model_args_dict)
    model = STR2MODEL["land_cover"](new_model_args)
    
    return model

def get_dataset_splits(add_geowiki: bool, add_nigeria: bool, geowiki_subset: str='world', S2_features_only: bool=False):

    # Dirty hack, we use landcovermapper class (only supporting lstm model) class to access model splits
    lancovermapper = get_model(add_geowiki, add_nigeria, geowiki_subset)

    train_dataset = lancovermapper.get_dataset(subset="training").as_array(flatten_x=True, S2_features_only=S2_features_only)
    val_dataset = lancovermapper.get_dataset(subset="validation", normalizing_dict=lancovermapper.normalizing_dict).as_array(flatten_x=True, S2_features_only=S2_features_only)
    #test_dataset = lancovermapper.get_dataset(subset="validation", normalizing_dict=lancovermapper.normalizing_dict, evaluating=True).as_array(flatten_x=True, S2_features_only=S2_features_only) 
    test_dataset = lancovermapper.get_dataset(subset="testing", normalizing_dict=lancovermapper.normalizing_dict).as_array(flatten_x=True, S2_features_only=S2_features_only)

    return (
        train_dataset,
        val_dataset,
        test_dataset
    )

# countries_subset = ['Nigeria']
# geowiki_dataset = GeowikiCropHarvestDataset(root=DATA_DIR/"cropharvest", countries_subset=countries_subset)
# geowiki_train, geowiki_val = geowiki_dataset.train_val_split(geowiki_dataset)

def get_dataset(subset: str, normalizing_dict: Optional[Dict] = None, evaluating: bool = False, add_geowiki: bool=True, add_nigeria: bool=True) -> LandTypeClassificationDataset:       
    # Geowiki
    geowiki_dataset = None
    if add_geowiki:
        # Define split
        if subset == 'training':
            geowiki_dataset = geowiki_train
        elif subset == 'validation':
            geowiki_dataset = geowiki_val

    # Nigeria
    nigeria_root_path = DATA_DIR / 'features' / 'nigeria-cropharvest'
    nigeria_dataset = None
    if add_nigeria or (subset == 'testing') or (subset == "validation" and evaluating):
        # We want to define Nigeria dataset in the following not mutually exclusive cases: 1) for training (i.e. when add_nigeria=True);
        # 2) for testing with test split; 3) for testing with validation split during development (i.e. evaluating=True).
        nigeria_dataset = NigeriaCropHarvestDataset(nigeria_root_path, split=subset)      

    return LandTypeClassificationDataset(
        subset=subset,
        include_geowiki=add_geowiki,
        include_nigeria=add_nigeria,
        evaluating=evaluating,
        geowiki_dataset=geowiki_dataset,
        nigeria_dataset=nigeria_dataset,
        normalizing_dict=normalizing_dict,
    )

# Train Random Forest model

 Evaluate on Nigeria test set

### 1. Training only on geowiki

### 1.1 Geowiki world

add_geowiki = True
add_nigeria = False
geowiki_subset = 'world' # if add_geowiki=False this will be ignored
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### Only with S2 data

(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset, S2_features_only=True)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 1.2 Geowiki neighbours

add_geowiki = True
add_nigeria = False
geowiki_subset = 'neighbours1' # if add_geowiki=False this will be ignored
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### Only with S2 data

(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset, S2_features_only=True)
print('')
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 1.3 Geowiki nigeria

add_geowiki = True
add_nigeria = False
geowiki_subset = 'nigeria' # if add_geowiki=False this will be ignored
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### Only with S2 data

(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset, S2_features_only=True)
print('')
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 2.1 Training only with Nigeria train set

add_geowiki = False
add_nigeria = True
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### Only with S2 data

(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset, S2_features_only=True) # geowiki subset doesn't matter here
print('')
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 3. Training using Nigeria validation set as well

add_geowiki = False
add_nigeria = True
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria)

# Normalizing dicts will be the same anyways as they were calculated with all 
X_train_val = np.concatenate((X_train, X_val))
y_train_val = np.concatenate((y_train, y_val))
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train_val, y_train_val)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

Using validation data for training seems to worsen the results a bit!

### 4. Train on geowiki and nigeria combined

### 4.1 Geowiki world

add_geowiki = True
add_nigeria = True
geowiki_subset = 'world'
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### Only with S2 data

(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset, S2_features_only=True)
print('')
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 4.2 Geowiki neighbours

add_geowiki = True
add_nigeria = True
geowiki_subset = 'neighbours1'
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset=geowiki_subset)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### Only with S2 data

(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset, S2_features_only=True)
print('')
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 4.3 Geowiki Nigeria

add_geowiki = True
add_nigeria = True
geowiki_subset = 'nigeria'
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset=geowiki_subset)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### Only with S2 data

(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset, S2_features_only=True)
print('')
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 5. Train on geowiki (nigeria subset -> best) and nigeria combined (S2 channels only) 

add_geowiki = True
add_nigeria = True
geowiki_subset = 'nigeria'
S2_features_only = True
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset=geowiki_subset, S2_features_only=S2_features_only)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 6.1 Train on Geowiki Nigeria only --> this shows the value of the hand-labelled data

add_geowiki = True
add_nigeria = False
geowiki_subset = 'nigeria'
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset=geowiki_subset)

rf = RandomForestClassifier(random_state=0)
rf.fit(X_train, y_train)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 6.2 Train on Geowiki Nigeria only (with validation set for training too)

X_train_val = np.concatenate((X_train, X_val))
y_train_val = np.concatenate((y_train, y_val))
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train_val, y_train_val)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd

### 7. Train on geowiki and nigeria combined + validation set

add_geowiki = True
add_nigeria = True
geowiki_subset = 'nigeria'
(X_train, y_train), (X_val, y_val), (X_test, y_test) = get_dataset_splits(add_geowiki, add_nigeria, geowiki_subset=geowiki_subset)

X_train_val = np.concatenate((X_train, X_val))
y_train_val = np.concatenate((y_train, y_val))
rf = RandomForestClassifier(random_state=0)
rf.fit(X_train_val, y_train_val)
#y_pred = rf.predict(X_test)
y_probs = rf.predict_proba(X_test)[:, 1] # for roc_auc_score
metrics = get_metrics(y_test, y_probs) # results are a bit better than in notebook 15 and 17 possibly due to correction in normalization of test set with training set nd
