==> notebooks/20_cropharvest_sentinel1_verify_nigeria_country_downloads.ipynb <==
from itertools import groupby
import os
from pathlib import Path
import re

import pandas as pd

## Check by index

file_info_path = Path('/media/Elements-12TB/satellite_images/nigeria/raw/nigeria-full-country-2020/gdrive_test_eo_data_files.csv')
os.system(f"ls -v {file_info_path.parent}/*.tif > filepaths.txt")

with open('filepaths.txt', 'r') as f:
    data = f.read().splitlines()

data = [Path(path).stem for path in data]
len(data)

counter = {}
for k, g in groupby(data, key=lambda x:re.search('(\d+)-', x).group(1)):
    counter[int(k)] = len(list(g))

### Completely missing

l1 = list(counter.keys())
l2 = list(range(1, 3564))
completely_missing = sorted(list(set(l2) - set(l1)))
len(completely_missing)

with open('missing_file_indexes.txt', 'w') as file:
    file.write('\n'.join(str(i) for i in completely_missing))

### Partially missing

missing_counter = {index: 4-count for index, count in counter.items()}
partially_missing = {index: count for index, count in missing_counter.items() if count > 0}
partially_missing

## Check by download list

df = pd.read_csv(file_info_path)

# Reconstruct paths and check if they all exist
paths = df['name'].apply(lambda x : file_info_path.parent / x.split('/')[-1]).to_list()
all([path.exists() for path in paths])
