==> notebooks/27_cropharvest_data_statistics.ipynb <==
import sys
from argparse import ArgumentParser, Namespace
from pathlib import Path

sys.path.append('..')

from src.models import STR2MODEL

DATA_DIR = Path("../data")

def get_model(add_geowiki: bool, add_nigeria: bool, geowiki_subset: str):
    parser = ArgumentParser()
    parser.add_argument("--max_epochs", type=int, default=100)
    parser.add_argument("--patience", type=int, default=10)
    parser.add_argument("--gpus", type=int, default=0)
    parser.add_argument("--wandb", default=False, action="store_true")

    model_args = STR2MODEL["land_cover"].add_model_specific_args(parser).parse_args(args=[])
    new_model_args_dict = vars(model_args)

    # SET MODIFICATIONS TO DEFAULT MODEL ARGUMENTS:
    new_model_args_dict['add_geowiki'] = add_geowiki
    new_model_args_dict['add_nigeria'] = add_nigeria
    new_model_args_dict['geowiki_subset'] =  geowiki_subset # 'nigeria', 'neighbours1'

    new_model_args = Namespace(**new_model_args_dict)
    model = STR2MODEL["land_cover"](new_model_args)
    
    return model


def set_statistics(df):
    total = len(df)
    value_counts = df.is_crop.value_counts()
    cropland_ratio = df.is_crop.sum() / total

    print(f"Total points: {total}")
    print(f"Cropland count: {value_counts[1]}")
    print(f"Non-cropland count: {value_counts[0]}")
    print(f"Cropland ratio: {cropland_ratio:.3f}")
    

## Geowiki

### Geowiki world

add_geowiki = True
add_nigeria = False
geowiki_subset = 'world'
landcovermapper = get_model(add_geowiki, add_nigeria, geowiki_subset)
df = landcovermapper.geowiki_dataset.labels
df.is_crop.value_counts()

set_statistics(df)

set_statistics(landcovermapper.geowiki_train.labels)

set_statistics(landcovermapper.geowiki_val.labels)

### Geowiki neighbours

add_geowiki = True
add_nigeria = False
geowiki_subset = 'neighbours1'
landcovermapper = get_model(add_geowiki, add_nigeria, geowiki_subset)
df = landcovermapper.geowiki_dataset.labels
df.is_crop.value_counts()

set_statistics(df)

set_statistics(landcovermapper.geowiki_train.labels)

set_statistics(landcovermapper.geowiki_val.labels)

### Geowiki nigeria

add_geowiki = True
add_nigeria = False
geowiki_subset = 'nigeria'
landcovermapper = get_model(add_geowiki, add_nigeria, geowiki_subset)
df = landcovermapper.geowiki_dataset.labels
df.is_crop.value_counts()

set_statistics(df)

set_statistics(landcovermapper.geowiki_train.labels)

set_statistics(landcovermapper.geowiki_val.labels)

## Nigeria

add_geowiki = False
add_nigeria = True
landcovermapper = get_model(add_geowiki, add_nigeria, geowiki_subset)

train_dataset = landcovermapper.get_dataset(subset="training")
val_dataset = landcovermapper.get_dataset(subset="validation", normalizing_dict=landcovermapper.normalizing_dict)
test_dataset = landcovermapper.get_dataset(subset="testing", normalizing_dict=landcovermapper.normalizing_dict)

len(train_dataset) + len(val_dataset) + len(test_dataset)

# Cropland points
train_dataset.datasets['nigeria'].is_crop.sum() + val_dataset.datasets['nigeria'].is_crop.sum() + test_dataset.datasets['nigeria'].is_crop.sum()

# Cropland ratio
round((train_dataset.datasets['nigeria'].is_crop.sum() + val_dataset.datasets['nigeria'].is_crop.sum() + test_dataset.datasets['nigeria'].is_crop.sum()) / (train_dataset.datasets['nigeria'].is_crop.shape[0] + val_dataset.datasets['nigeria'].is_crop.shape[0] + test_dataset.datasets['nigeria'].is_crop.shape[0]), 3)

set_statistics(train_dataset.datasets['nigeria'])

set_statistics(val_dataset.datasets['nigeria'])

set_statistics(test_dataset.datasets['nigeria'])
